Dev

See merge request si/esup-signature!1939

+++ src/main/resources/static/js/modules/ui/signrequests/WorkspacePdf.js
@@ -75,8 +75,10 @@ export class WorkspacePdf {
             if(this.wsTabs.length) {
                 this.autocollapse();
                 let self = this;
-                $(window).on('resize', function () {
-                    self.autocollapse();
+                $(window).on('resize', function (e) {
+                    if(e.target.tagName == null) {
+                        self.autocollapse();
+                    }
                 });
                 if(this.secondTools.children().length > 0) {
                     this.workspace.css("margin-top", "216px");
@@ -208,9 +210,9 @@ export class WorkspacePdf {
                 signSpaceDiv.on("click", e => this.addSign(i));
                 if(currentSignRequestParams.ready == null || !currentSignRequestParams.ready) {
                     if(this.signType !== "visa") {
-                        signSpaceDiv.html("Cliquez ici pour insérer votre signature<br>" + currentSignRequestParams.comment);
+                        signSpaceDiv.html("Cliquez ici pour insérer votre signature");
                     } else {
-                        signSpaceDiv.html("Cliquez ici pour insérer votre visa<br>" + currentSignRequestParams.comment);
+                        signSpaceDiv.html("Cliquez ici pour insérer votre visa");
                     }
                 }
                 if (currentSignRequestParams.ready) {
@@ -580,24 +582,26 @@ export class WorkspacePdf {
                         signDiv.css("font-size", 12 * self.pdfViewer.scale);
                     }
                     spotDiv.unbind('mouseup');
-                    spotDiv.on('mouseup', function (e) {
-                        e.stopPropagation();
-                        bootbox.confirm("Supprimer cet emplacement de signature ?", function (result) {
-                            if (result) {
-                                let url = "/ws-secure/global/delete-comment/" + self.signRequestId + "/" + spot.id + "?" + self.csrf.parameterName + "=" + self.csrf.token;
-                                if(self.currentSignType === "form") {
-                                    url = "/admin/forms/delete-spot/" + self.formId + "/" + spot.id + "?" + self.csrf.parameterName + "=" + self.csrf.token;
-                                }
-                                $.ajax({
-                                    method: 'DELETE',
-                                    url: url,
-                                    success: function () {
-                                        spotDiv.remove();
+                    if(signDiv.attr("data-es-delete")) {
+                        spotDiv.on('mouseup', function (e) {
+                            e.stopPropagation();
+                            bootbox.confirm("Supprimer cet emplacement de signature ?", function (result) {
+                                if (result) {
+                                    let url = "/ws-secure/global/delete-comment/" + self.signRequestId + "/" + spot.id + "?" + self.csrf.parameterName + "=" + self.csrf.token;
+                                    if (self.currentSignType === "form") {
+                                        url = "/admin/forms/delete-spot/" + self.formId + "/" + spot.id + "?" + self.csrf.parameterName + "=" + self.csrf.token;
                                     }
-                                });
-                            }
+                                    $.ajax({
+                                        method: 'DELETE',
+                                        url: url,
+                                        success: function () {
+                                            spotDiv.remove();
+                                        }
+                                    });
+                                }
+                            });
                         });
-                    });
+                    }
                 }
                 index++;
             }
@@ -775,6 +779,7 @@ export class WorkspacePdf {
         localStorage.setItem('mode', 'comment');
         $("#postitHelp").remove();
         this.disableAllModes();
+        $("#changeMode1").attr("checked", true);
         $("#postit").removeClass("d-none");
         $("#commentHelp").removeClass("d-none");
         this.mode = 'comment';
@@ -807,6 +812,7 @@ export class WorkspacePdf {
         console.info("enable sign mode");
         localStorage.setItem('mode', 'sign');
         this.disableAllModes();
+        $("#changeMode2").attr("checked", true);
         this.mode = 'sign';
         this.signPosition.pointItEnable = false;
         if (this.status === 'pending') {
@@ -1011,27 +1017,15 @@ export class WorkspacePdf {
         //     text: 'Lecture',
         //     value: 'read'
         // });
-
-        if($("#changeMode").length) {
-            this.changeModeSelector = new SlimSelect({
-                select: '#changeMode',
-                settings: {
-                    showSearch: false,
-                    valuesUseText: false,
-                },
-                events: {
-                    afterChange: (val) => {
-                        this.changeMode(val)
-                    }
-                },
-            });
-            this.changeModeSelector.setData(data);
-        }
+        let self = this;
+        $("#changeMode1").on("click", function(e) {
+            self.changeMode("comment");
+        });
+        $("#changeMode2").on("click", function(e) {
+            self.changeMode("sign");
+        });
         if(this.changeModeSelector != null) {
-            if(this.signable) {
-                this.changeModeSelector.setSelected("sign");
-            } else {
-                this.changeModeSelector.setSelected("comment");
+            if(!this.signable) {
                 this.enableCommentMode();
             }
         }
@@ -1039,7 +1033,7 @@ export class WorkspacePdf {
 
     changeMode(e) {
         if(this.ready) {
-            let mode = e[0].value;
+            let mode = e;
             console.info("change mode to : " + mode);
             if (mode === "sign" && this.signable) {
                 this.enableSignMode();
